refactor: 어드민 페이지 섹션 분리 및 state 사용하는 곳으로 컴포넌트 디테일 분리 (미완성)
